--check before destructing dest *ptr that it isn't the same as src *ptr
Not sure if this could even happen (a situation could arise where two different instances could have their _data arrays both be pointers to the same memory location, but just in case.

+++ src/esp/core/Configuration.cpp
@@ -186,7 +186,7 @@ void ConfigValue::deleteCurrentValue() {
 }
 
 ConfigValue& ConfigValue::operator=(const ConfigValue& otr) {
-  if (this != &otr) {
+  if ((this != &otr) && isSafeToDeconstruct(otr)) {
     deleteCurrentValue();
     copyValueFrom(otr);
   }
@@ -194,8 +194,11 @@ ConfigValue& ConfigValue::operator=(const ConfigValue& otr) {
 }  // ConfigValue::operator=
 
 ConfigValue& ConfigValue::operator=(ConfigValue&& otr) noexcept {
-  deleteCurrentValue();
+  if (isSafeToDeconstruct(otr)) {
+    deleteCurrentValue();
+  }
   moveValueFrom(std::move(otr));
+
   return *this;
 }  // ConfigValue::operator=
